Fix: Error accessing cookies in fusionauth. If you are using SSR you must configure the SDK with a cookie adapter
When using the `next-client-cookies` as per documentation, the `nextCookieAdapter` is not passed as it's not included in the config. This causes the coreSDK to fail to grab the cookies. 

As the config object is using the `SDKConfig` and `FusionAuthProviderConfig` is expected, I've change the function parameter from `config` to `props` to pass the correct `FusionAuthProviderConfig` object forward.